requirements.txt
accelerate==0.24.1
beautifulsoup4==4.12.2
boto3==1.28.75
botocore==1.31.75
bpemb==0.3.4
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
conllu==4.5.3
contourpy==1.1.1
cycler==0.12.1
Deprecated==1.2.14
diskcache==5.6.3
filelock==3.12.4
flair==0.13.0
fonttools==4.43.1
fsspec==2023.10.0
ftfy==6.1.1
gdown==4.7.1
gensim==4.3.2
huggingface-hub==0.17.3
idna==3.4
Janome==0.5.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
langdetect==1.0.9
llama_cpp_python==0.2.11
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
more-itertools==10.1.0
mpld3==0.5.9
mpmath==1.3.0
multipledispatch==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
pptree==3.1
protobuf==4.24.4
psutil==5.9.6
pyparsing==3.1.1
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
pytorch_revgrad==0.2.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
s3transfer==0.7.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
segtok==1.5.11
semver==3.0.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
sqlitedict==2.1.0
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
transformer-smaller-training-vocab==0.3.2
transformers==4.34.1
typing_extensions==4.8.0
urllib3==1.26.18
wcwidth==0.2.9
Wikipedia-API==0.6.0
wrapt==1.15.0
